shadow-dom/shadow-trees/nested-shadow-trees/do-not-apply-author-style.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Shadow DOM - do not apply author style</title>
  <link rel="author" title="thujikun" href="mailto:tsuji.0107@gmail.com" />
  <link rel="help" href="http://www.w3.org/TR/shadow-dom/" />
  <meta name="assert" content="Nested tree: attribute bool applyAuthorStyle attribute. If true, the author styles are not applied">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="../../testcommon.js"></script>
 </head>
 <body>
    <h1>nested shadow trees level 2</h1>

    <style>
      #level2-red {
        width: 100px;
        height: 100px;
        background: #ff0000;
      }
    </style>
    <div id="level0"></div>

    <div id="log"></div>
    <template id="level1-template">
      <style>
        #level2-blue {
          width: 100px;
          height: 100px;
          background: #0000ff;
        }
      </style>

      <div id="level1">
        <div id="level2-root"></div>
      </div>

    </template>
    <template id="level2-template">
      <style>
        #level2-green {
          width: 100px;
          height: 100px;
          background: #00ff00;
        }
      </style>
      <div id="level2-red">
      </div>
      <div id="level2-blue">
      </div>
      <div id="level2-green">
      </div>
    </template>
    <script>
        test(function() {
            var level0 = document.getElementById('level0'),
                shadowLevel1 = createSR(level0),
                shadowLevel2,
                level1 = document.getElementById('level1-template').content,
                level2 = document.getElementById('level2-template').content,
                getComputedStyle = document.defaultView.getComputedStyle;

            shadowLevel1.appendChild(level1);

            shadowLevel2 = shadowLevel1.getElementById('level2-root').webkitCreateShadowRoot();

            shadowLevel2.appendChild(level2);

            assert_not_equals(getComputedStyle(shadowLevel2.querySelector('#level2-red')).backgroundColor, 'rgb(255, 0, 0)');
            assert_not_equals(getComputedStyle(shadowLevel2.querySelector('#level2-blue')).backgroundColor, 'rgb(0, 0, 255)');
            assert_equals(getComputedStyle(shadowLevel2.querySelector('#level2-green')).backgroundColor, 'rgb(0, 255, 0)');

        }, 'do not apply author style on nested tree');
    </script>
 </body>
</html>
